Refactor OperationRow
