+++ shared/crypto/operations-list/index.stories.tsx
@@ -1,24 +1,14 @@
 import * as React from 'react'
 import * as Sb from '../../stories/storybook'
 import * as Constants from '../../constants/crypto'
-// import * as Types from '../../constants/types/crypto'
 import OperationsList from '.'
-
-const onSelect = Sb.action('onSelect')
-
-const provider = Sb.createPropProvider({
-  OperationRow: props => ({
-    isSelected: props.isSelected || false,
-    onSelect: () => onSelect(props.tab),
-    tab: props.tab,
-    title: props.title,
-  }),
-})
 
 const load = () => {
   Sb.storiesOf('Crypto/Operations List', module)
-    .addDecorator(provider)
-    .add('Default', () => <OperationsList routeSelected={Constants.encryptTab} />)
+    .add('Select - Encrypt', () => <OperationsList routeSelected={Constants.encryptTab} />)
+    .add('Select - Decrypt', () => <OperationsList routeSelected={Constants.decryptTab} />)
+    .add('Select - Sign', () => <OperationsList routeSelected={Constants.signTab} />)
+    .add('Select - Verify', () => <OperationsList routeSelected={Constants.verifyTab} />)
 }
 
 export default load
